Deterministic a parameter
