--- torchsynth/__init__.py
@@ -1,3 +1,0 @@
-from . import reproducibility
-
-reproducibility.check_for_determinism()
